+++ src/solana/rpc/core.py
@@ -14,6 +14,7 @@ from based58 import b58decode, b58encode
 
 from solana.blockhash import Blockhash, BlockhashCache
 from solana.keypair import Keypair
+from solana.message import Message
 from solana.publickey import PublicKey
 from solana.rpc import types
 from solana.transaction import Transaction
@@ -31,6 +32,14 @@ class RPCNoResultException(Exception):
 
 class UnconfirmedTxError(Exception):
     """Raise when confirming a transaction times out."""
+
+
+class TransactionExpiredBlockheightExceededError(Exception):
+    """Raise when confirming an expired transaction that exceeded the blockheight."""
+
+
+class TransactionUncompiledError(Exception):
+    """Raise when transaction is not compiled to a message."""
 
 
 class _ClientCore:  # pylint: disable=too-few-public-methods
@@ -153,11 +162,11 @@ class _ClientCore:  # pylint: disable=too-few-public-methods
         limit: Optional[int],
         commitment: Optional[Commitment],
     ) -> Tuple[types.RPCMethod, str, Dict[str, Union[int, str, Commitment]]]:
-        # warn(
-        #     "solana.rpc.api.getConfirmedSignaturesForAddress2 is deprecated, "
-        #     "please use solana.rpc.api.getSignaturesForAddress",
-        #     category=DeprecationWarning,
-        # )
+        warn(
+            "solana.rpc.api.getConfirmedSignaturesForAddress2 is deprecated, "
+            "please use solana.rpc.api.getSignaturesForAddress",
+            category=DeprecationWarning,
+        )
         opts = self._get_signature_for_address_config_arg(before, until, limit, commitment)
         account = self._get_signature_for_address_account_arg(account)
         return types.RPCMethod("getConfirmedSignaturesForAddress2"), account, opts
@@ -239,6 +248,16 @@ class _ClientCore:  # pylint: disable=too-few-public-methods
         return (
             types.RPCMethod("getFeeCalculatorForBlockhash"),
             blockhash,
+            {self._comm_key: commitment or self._commitment},
+        )
+
+    def _get_fee_for_message_args(
+        self, message: Message, commitment: Optional[Commitment]
+    ) -> Tuple[types.RPCMethod, str, Dict[str, Commitment]]:
+        raw_message = b64encode(message.serialize()).decode("utf-8")
+        return (
+            types.RPCMethod("getFeeForMessage"),
+            raw_message,
             {self._comm_key: commitment or self._commitment},
         )
 
@@ -596,8 +615,8 @@ class _ClientCore:  # pylint: disable=too-few-public-methods
     @staticmethod
     def _send_raw_transaction_post_send_args(
         resp: types.RPCResponse, opts: types.TxOpts
-    ) -> Tuple[types.RPCResponse, Commitment]:
-        return resp, opts.preflight_commitment
+    ) -> Tuple[types.RPCResponse, Commitment, Optional[int]]:
+        return resp, opts.preflight_commitment, opts.last_valid_block_height
 
     def _simulate_transaction_args(
         self,