+++ tests/integration/test_memo.py
@@ -8,12 +8,15 @@ from solana.transaction import Transaction
 from spl.memo.constants import MEMO_PROGRAM_ID
 from spl.memo.instructions import MemoParams, create_memo
 
-from .utils import assert_valid_response
+from .utils import AIRDROP_AMOUNT, assert_valid_response
 
 
 @pytest.mark.integration
 def test_send_memo_in_transaction(stubbed_sender: Keypair, test_http_client: Client):
     """Test sending a memo instruction to localnet."""
+    airdrop_resp = test_http_client.request_airdrop(stubbed_sender.public_key, AIRDROP_AMOUNT)
+    assert_valid_response(airdrop_resp)
+    test_http_client.confirm_transaction(airdrop_resp["result"])
     raw_message = "test"
     message = bytes(raw_message, encoding="utf8")
     # Create memo params
@@ -29,7 +32,8 @@ def test_send_memo_in_transaction(stubbed_sender: Keypair, test_http_client: Cli
     resp = test_http_client.send_transaction(transfer_tx, stubbed_sender)
     assert_valid_response(resp)
     txn_id = resp["result"]
-    test_http_client.confirm_transaction(txn_id)
+    # Txn needs to be finalized in order to parse the logs.
+    test_http_client.confirm_transaction(txn_id, commitment=Finalized)
     resp2 = test_http_client.get_transaction(txn_id, commitment=Finalized, encoding="jsonParsed")
     log_message = resp2["result"]["meta"]["logMessages"][2].split('"')
     assert log_message[1] == raw_message