+++ tests/integration/utils.py
@@ -1,7 +1,7 @@
 """Integration test utils."""
 
-from solana.rpc.types import RPCResponse, TxOpts
 from solana.rpc.commitment import Processed
+from solana.rpc.types import RPCResponse, TxOpts
 
 AIRDROP_AMOUNT = 10_000_000_000
 
@@ -10,7 +10,7 @@ def assert_valid_response(resp: RPCResponse):
     """Assert valid RPCResponse."""
     assert resp["jsonrpc"] == "2.0"
     assert resp["id"]
-    assert resp["result"]
+    assert resp["result"] is not None
 
 
 def compare_responses_without_ids(left: RPCResponse, right: RPCResponse) -> None: